requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
defusedxml==0.7.1
Django==4.0
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.13.1
docopt==0.6.2
entrypoints==0.3
firebase-admin==5.2.0
fonttools==4.28.4
frozenlist==1.2.0
idna==3.3
imageio==2.13.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipython-genutils==0.2.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonmerge==1.8.0
jsonpickle==2.1.0
jsonschema==4.3.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
mistune==0.8.4
mpmath==1.2.1
multidict==5.2.0
nbclient==0.5.9
ndim==0.1.5
nest-asyncio==1.5.4
orthopy
packaging==21.3
pandocfilters==1.5.0
Pillow==8.4.0
pipreqs==0.4.11
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
Pygments==2.10.0
-e git+https://github.com/ecs-kev/pyleecan#egg=pyleecan&subdirectory=../../../gitsubmodules/pyleecan
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.3
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
s3transfer==0.5.0
scikit-learn==1.0.1
scooby==0.5.9
shiboken2==5.15.2
six==1.16.0
sqlparse==0.4.2
sympy==1.9
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.7
vtk==9.1.0
webencodings==0.5.1
wslink==1.3.0
XlsxWriter==3.0.2
yarg==0.1.9
yarl==1.7.2
zipp==3.6.0
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
celery==5.2.7
redis==4.3.3
django-ses==3.1.0